cleanup: move member infos computation in Member model
