+++ lib/api.js
@@ -5,8 +5,9 @@ import mongo from './util/mongo.js'
 import {sendMail} from './util/sendmail.js'
 import renderFinAbonnement from './emails/fin-abonnement.js'
 import renderPlusDeTickets from './emails/plus-de-tickets.js'
-import {addTickets, addAbo, addMembership, getUserIdByEmail, findOrCreateRelatedUserId, updateBalance, computeUserStats, computeMembershipOk, computeLastMembership, syncUser} from './models.js'
+import {addTickets, addAbo, addMembership, getUserIdByEmail, findOrCreateRelatedUserId, updateBalance, syncUser} from './models.js'
 
+import * as Member from './models/member.js'
 import * as Presence from './models/presence.js'
 
 function formatDate(date) {
@@ -30,8 +31,8 @@ export async function coworkersNow(req, res) {
 }
 
 export async function getUserStats(req, res) {
-  const userStats = computeUserStats(req.rawUser)
-  res.send(userStats)
+  const member = await Member.getMemberById(req.rawUser._id)
+  res.send(member)
 }
 
 export async function getUserPresences(req, res) {
@@ -331,28 +332,8 @@ export async function getCurrentUsers(req, res) {
   }
 
   const delay = userDefinedDelay || 10
-  const dateToCompare = sub(new Date(), {minutes: delay}).toISOString()
-
-  const users = await mongo.db.collection('users')
-    .find({'profile.heartbeat': {$gt: dateToCompare}})
-    .project({'profile.presences': 0, 'profile.tickets': 0})
-    .toArray()
-
-  res.send(users.map(user => {
-    const lastMembership = computeLastMembership(user.profile.memberships)
-    const membershipOk = computeMembershipOk(lastMembership)
-
-    return {
-      _id: user._id,
-      wpUserId: user.wpUserId,
-      firstName: user.profile.firstName,
-      lastName: user.profile.lastName,
-      balance: user.profile.balance,
-      birthDate: user.profile.birthDate,
-      lastMembership,
-      membershipOk
-    }
-  }))
+  const currentMembers = await Member.getCurrentMembers(delay)
+  res.send(currentMembers)
 }
 
 export async function getVotingCoworkers(req, res) {
@@ -368,14 +349,8 @@ export async function getVotingCoworkers(req, res) {
 
   const minActivity = req.minActivity || 20
 
-  const users = await mongo.db.collection('users').find({}).toArray()
-  res.send(chain(users)
-    .map(u => computeUserStats(u))
-    .filter(u => u.activity >= minActivity)
-    .map(u => pick(u, 'firstName', 'lastName', 'email', 'activity', 'lastMembership', 'balance'))
-    .sortBy(u => -u.activity)
-    .value()
-  )
+  const votingMembers = await Member.getVotingMembers(minActivity)
+  res.send(votingMembers)
 }
 
 export function convertDate(frDate) {
