+++ lib/models.js
@@ -1,6 +1,6 @@
 /* eslint-disable unicorn/no-array-reduce */
 import {add, sub} from 'date-fns'
-import {chain, range, minBy, maxBy, sumBy, sortBy} from 'lodash-es'
+import {range, minBy, sumBy} from 'lodash-es'
 import {customAlphabet} from 'nanoid'
 
 import mongo from './util/mongo.js'
@@ -204,77 +204,6 @@ export async function updateBalance(userId) {
   )
 
   return balance
-}
-
-export function computeLastMembership(memberships) {
-  const lastMembership = maxBy(memberships, 'purchaseDate')
-
-  if (lastMembership) {
-    const purchaseMonth = Number.parseInt(lastMembership.purchaseDate.slice(5, 7), 10)
-    const purchaseYear = Number.parseInt(lastMembership.purchaseDate.slice(0, 4), 10)
-    return String(purchaseMonth >= 11 ? purchaseYear + 1 : purchaseYear)
-  }
-}
-
-export function computeMembershipOk(lastMembership) {
-  const now = new Date()
-  const currentYear = now.toISOString().slice(0, 4)
-
-  return lastMembership && lastMembership >= currentYear
-}
-
-export function computeUserStats(user) {
-  const today = (new Date()).toISOString().slice(0, 10)
-
-  const userStats = {
-    firstName: user.profile.firstName,
-    lastName: user.profile.lastName,
-    email: user.emails[0].address
-  }
-  userStats.balance = user.profile.balance
-
-  const lastAbo = maxBy(user.profile.abos, 'aboStart')
-
-  if (lastAbo) {
-    const lastAboEnd = sub(add(new Date(lastAbo.aboStart), {months: 1}), {days: 1})
-      .toISOString().slice(0, 10)
-
-    if (today <= lastAboEnd) {
-      userStats.lastAboEnd = lastAboEnd
-    }
-  }
-
-  userStats.lastMembership = computeLastMembership(user.profile.memberships)
-  userStats.membershipOk = computeMembershipOk(userStats.lastMembership)
-
-  const sixMonthsAgo = sub(new Date(), {months: 6}).toISOString().slice(0, 10)
-
-  const sixMonthsActivity = chain(user.profile.presences)
-    .filter(p => p.date >= sixMonthsAgo)
-    .sumBy('amount')
-    .value()
-
-  userStats.activity = sixMonthsActivity
-  userStats.activeUser = sixMonthsActivity >= 20
-
-  const totalActivity = sumBy(user.profile.presences, 'amount')
-
-  userStats.totalActivity = totalActivity
-  userStats.presencesConso = totalActivity
-  userStats.presencesJours = user.profile.presences.length
-  userStats.trustedUser = totalActivity >= 10
-
-  const abos = sortBy(user.profile.abos, 'aboStart')
-    .map(abo => {
-      const {aboStart, purchaseDate} = abo
-      const aboEnd = sub(add(new Date(aboStart), {months: 1}), {days: 1}).toISOString().slice(0, 10)
-      const current = today >= aboStart && today <= aboEnd
-      return {purchaseDate, aboStart, aboEnd, current}
-    })
-    .reverse()
-
-  userStats.abos = abos
-  return userStats
 }
 
 export async function getUsers() {
